app/requirements.txt
Django==5.0.4
psycopg[binary]==3.1.18
python-dotenv==1.0.1
split-settings==1.0.0
uWSGI==2.0.26
django-debug-toolbar==4.4.6
django-cors-headers==4.4.0